ClustForOpt.jl Documentation

ClustForOpt.jl Documentation

Example

ClustForOpt.get_cep_variable_set — Function.

get_cep_variable_set(variable::OptVariable,num_index_set::Int)

Get the variable set from the specific variable and the num_index_set like 1

source

get_cep_variable_set(scenario::Scenario,var_name::String,num_index_set::Int)

Get the variable set from the specific Scenario by indicating the var_name e.g. "COST" and the num_index_set like 1

source